(#5192) - improve test coverage
