+++ packages/pouchdb-adapter-fruitdown/src/index.js
@@ -1,5 +1,4 @@
 import CoreLevelPouch from 'pouchdb-adapter-leveldb-core';
-import { toPromise } from 'pouchdb-utils';
 import { extend } from 'js-extend';
 
 import fruitdown from 'fruitdown';
@@ -18,17 +17,6 @@ FruitDownPouch.valid = function () {
 };
 FruitDownPouch.use_prefix = true;
 
-FruitDownPouch.destroy = toPromise(function (name, opts, callback) {
-  if (typeof opts === 'function') {
-    callback = opts;
-    opts = {};
-  }
-  var _opts = extend({
-    db: fruitdown
-  }, opts);
-
-  return fruitdown.destroy(name, _opts, callback);
-});
 export default function (PouchDB) {
   PouchDB.adapter('fruitdown', FruitDownPouch, true);
 }
