+++ packages/pouchdb-adapter-leveldb/package.json
@@ -15,7 +15,6 @@
   "dependencies": {
     "js-extend": "1.0.1",
     "leveldown": "1.4.6",
-    "pouchdb-adapter-leveldb-core": "5.4.0-prerelease",
-    "pouchdb-utils": "5.4.0-prerelease"
+    "pouchdb-adapter-leveldb-core": "5.4.0-prerelease"
   }
 }
