+++ packages/pouchdb-binary-utils/package.json
@@ -19,6 +19,7 @@
     "./src/arrayBufferToBinaryString.js": "./src/arrayBufferToBinaryString-browser.js",
     "./src/base64.js": "./src/base64-browser.js",
     "./src/base64StringToBlobOrBuffer.js": "./src/base64StringToBlobOrBuffer-browser.js",
+    "./src/blob.js": "./src/blob-browser.js",
     "./src/binaryStringToBlobOrBuffer.js": "./src/binaryStringToBlobOrBuffer-browser.js",
     "./src/blobOrBufferToBase64.js": "./src/blobOrBufferToBase64-browser.js",
     "./src/readAsArrayBuffer.js": "./src/readAsArrayBuffer-browser.js"
